✨ Add support for gitlab and bitbucket

+++ etc/init.d/30-forge.sh
@@ -33,11 +33,11 @@ bin_name
 
 : "${FORGES_USER:=""}"
 
-# GitHub user to fetch keys from
-: "${FORGES_GITHUB_USER:="${TUNNEL_GITHUB_USER:-""}"}"
-
 # Environment file to load for reading defaults from.
 : "${FORGES_DEFAULTS:="${FORGES_ROOTDIR}/../${CODER_BIN}.env"}"
+
+# List of domains/forges that we support
+: "${FORGES_SUPPORTED:="github.com gist.github.com bitbucket.org gitlab.com"}"
 
 
 log_init FORGES
@@ -69,38 +69,44 @@ json2sha256() {
 }
 
 
-authorize_github() {
-  # Download fingerprints reported by GH API
-  _fingerprints=$(mktemp json.XXXXXX)
-  download https://api.github.com/meta "$_fingerprints"
+authorize_forges() {
+  if printf %s\\n "$FORGES_SUPPORTED" | grep -qF 'github.com'; then
+    # Download fingerprints reported by GH API
+    _fingerprints=$(mktemp json.XXXXXX)
+    download https://api.github.com/meta "$_fingerprints"
+  fi
 
-  for _domain in github.com gist.github.com; do
+  for _domain in $FORGES_SUPPORTED; do
     verbose "Adding %s keys to %s" "$_domain" "${HOME}/.ssh/known_hosts"
-    _gh_keys=$(mktemp known_hosts.XXXXXX)
-    ssh-keyscan "$_domain" | grep -vE -e '^#' -e '^\s+$' > "$_gh_keys"
-    for _crypto in RSA ECDSA ED25519; do
+    _forge_keys=$(mktemp known_hosts.XXXXXX)
+    ssh-keyscan "$_domain" | grep -vE -e '^#' -e '^\s+$' > "$_forge_keys"
+    for _crypto in RSA DSA ECDSA ED25519; do
       if grep -F "$_domain" "${HOME}/.ssh/known_hosts" | grep -qF "$(to_lower "${_crypto}")" ; then
         debug "%s %s key already in %s" "$_domain" "${_crypto}" "${HOME}/.ssh/known_hosts"
       else
         verbose "Adding %s %s key to %s" "$_domain" "${_crypto}" "${HOME}/.ssh/known_hosts"
-        # Compute the fingerprint of the key, see https://serverfault.com/a/701637
-        _reported=$(sshkey2sha256 "$_gh_keys" "$_crypto")
-        trace "sha256 sum as reported by the live server is: %s" "$_reported"
-        # Extract the official fingerprint from the GitHub API. See
-        # https://serverfault.com/a/701637
-        _official=$(json2sha256 "$_fingerprints" "$_crypto")
-        trace "GitHub's API reports sha256 sum should be: %s" "$_official"
-        # If they match, add the key to the known_hosts file
-        if [ "$_reported" = "$_official" ]; then
-          grep -F "$(to_lower "${_crypto}")" "$_gh_keys" >> "${HOME}/.ssh/known_hosts"
+        if printf %s\\n "$_domain" | grep -qF 'github.com'; then
+          # Compute the fingerprint of the key, see https://serverfault.com/a/701637
+          _reported=$(sshkey2sha256 "$_forge_keys" "$_crypto")
+          trace "sha256 sum as reported by the live server is: %s" "$_reported"
+          # Extract the official fingerprint from the GitHub API. See
+          # https://serverfault.com/a/701637
+          _official=$(json2sha256 "$_fingerprints" "$_crypto")
+          trace "GitHub's API reports sha256 sum should be: %s" "$_official"
+          # If they match, add the key to the known_hosts file
+          if [ "$_reported" = "$_official" ]; then
+            grep -F "$(to_lower "${_crypto}")" "$_forge_keys" >> "${HOME}/.ssh/known_hosts"
+          else
+            warn "Fingerprint for %s key does not match official fingerprint" "${_crypto}"
+          fi
         else
-          warn "Fingerprint for %s key does not match official fingerprint" "${_crypto}"
+          grep -F "$(to_lower "${_crypto}")" "$_forge_keys" >> "${HOME}/.ssh/known_hosts"
         fi
       fi
     done
-    rm -f "$_gh_keys"
+    rm -f "$_forge_keys"
   done
-  rm -f "$_fingerprints"
+  [ -n "${_fingerprints:-}" ] && rm -f "$_fingerprints"
 }
 
 
@@ -118,4 +124,4 @@ if ! [ -f "${HOME}/.ssh/known_hosts" ]; then
 fi
 ensure_ownership "${HOME}/.ssh/known_hosts"
 
-authorize_github
+authorize_forges
